Update for better compat with Live Player

+++ Hanselman.Portable/Views/MenuPage.xaml.cs
@@ -45,7 +45,7 @@ namespace Hanselman.Portable.Views
                     if(ListViewMenu.SelectedItem == null)
                         return;
 
-                    await this.root.NavigateAsync(((HomeMenuItem)e.SelectedItem).MenuType);
+                    await this.root.NavigateAsync((int)((HomeMenuItem)e.SelectedItem).MenuType);
                 };
         }
     }
